redirect to newly created series & move components to new folder
