+++ app/event_series/[id]/edit/page.tsx
@@ -1,4 +1,4 @@
-import EditEventSeriesForm from '@/components/EditEventSeriesForm';
+import EditEventSeriesForm from '@/app/event_series/components/edit_event_series_form';
 import prisma from '@/db';
 
 const EventSeriesEdit = async ({ params }: { params: { id: string } }) => {
